pyproject.toml
[tool.poetry]
name = "llm-service"
version = "0.1.0"
description = "gRPC service for language models running on Ollama"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "llm_service"}]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
protobuf = "^4.22.3"
requests = "^2.31.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pylint = "^2.17.4"
types-requests = "^2.31.0.1"
types-PyYAML = "^6.0.12.11"

[tool.poetry.scripts]
llm-service = "llm_service.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "grpc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "proto.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"